refactor: add DTOs
