+++ Backend/CampusTradeSystem/CampusTrade.API/infrastructure/Extensions/ClaimsPrincipalExtensions.cs
@@ -1,0 +1,16 @@
+using System.Security.Claims;
+
+namespace CampusTrade.API.Infrastructure.Extensions
+{
+    public static class ClaimsPrincipalExtensions
+    {
+        public static int GetUserId(this ClaimsPrincipal user)
+        {
+            var userIdClaim = user.FindFirst(ClaimTypes.NameIdentifier)?.Value
+                               ?? user.FindFirst("userId")?.Value;
+            if (string.IsNullOrEmpty(userIdClaim) || !int.TryParse(userIdClaim, out int userId))
+                throw new UnauthorizedAccessException("用户身份验证失败");
+            return userId;
+        }
+    }
+}
