+++ Backend/CampusTradeSystem/CampusTrade.Tests/UnitTests/Services/ProductServiceTests.cs
@@ -1,6 +1,8 @@
 using System.Linq.Expressions;
 using CampusTrade.API.Models.DTOs.Common;
 using CampusTrade.API.Models.DTOs.Product;
+using CampusTrade.API.Models.DTOs.Order;
+using CampusTrade.API.Models.DTOs.Payment;
 using CampusTrade.API.Models.Entities;
 using CampusTrade.API.Repositories.Interfaces;
 using CampusTrade.API.Services.File;
